requirements.txt
cffi==1.17.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.3
cycler==0.12.1
fonttools==4.55.4
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
rich==14.0.0
shellingham==1.5.4
six==1.17.0
typer==0.15.3
typing_extensions==4.13.2
